modified:   CMakeLists.txt 	modified:   README.md 	modified:   gpt2/cross_entropy.cpp 	modified:   gpt2/gpt2.cpp 	modified:   train.cpp

+++ gpt2/cross_entropy.cpp
@@ -23,7 +23,7 @@ float CrossEntropy::Forward(const VecBTV &probs, const Mati &targets) {
 
 	for (int b = 0; b < batchs; ++b) {
 		for (int t = 0; t < seq_len; ++t) {
-			losses[b](t) = -logf(probs[b](t, targets(b, t)));
+			losses[b](t) = -std::log(probs[b](t, targets(b, t)));
 			mean_loss += losses[b](t);
 		}
 	}
